Now Gal is writing to Rs232 without a Thread

+++ jemma.osgi.javagal/src/main/java/org/energy_home/jemma/javagal/layers/data/interfaces/IDataLayer.java
@@ -109,7 +109,7 @@ public interface IDataLayer extends IFrameCallback {
 	 * @throws Exception
 	 *             if a not ZGD error occurs.
 	 */
-	public BigInteger readExtAddress(long timeout, short shortAddress) throws GatewayException,
+	public BigInteger readExtAddress(long timeout, Integer shortAddress) throws GatewayException,
 			Exception;
 
 	/**
